+++ NEWS.md
@@ -1,5 +1,9 @@
 # version 1.1-1
 
+* add argument `include_non_intersected` to `st_interpolate_aw()`
+
+* add argument `weights`, for dasymetric mapping, to `st_interpolate_aw()` (experimental)
+
 * replace magrittr pipe `%>%` with base pipe `|>`, and hence require R >= 4.1.0
 
 # version 1.1-0